feat(codegen): tidy main dir exports
