+++ apps/connect/package.copy.json
@@ -1,7 +1,7 @@
 {
   "name": "@powerhousedao/connect",
   "productName": "Powerhouse-Connect",
-  "version": "5.1.0-dev.18",
+  "version": "5.1.0-dev.20",
   "description": "Powerhouse Connect",
   "main": "dist/index.html",
   "type": "module",
