+++ packages/codegen/package.json
@@ -11,8 +11,16 @@
     "type": "git",
     "url": "https://github.com/powerhouse-inc/powerhouse"
   },
-  "exports": "./dist/index.js",
-  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    },
+    "./ts-morph": {
+      "types": "./dist/src/ts-morph-utils/index.d.ts",
+      "default": "./dist/src/ts-morph-utils/index.js"
+    }
+  },
   "scripts": {
     "tsc": "tsc",
     "lint": "eslint",
